requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
dj-database-url==1.2.0
Django==4.1.4
django-crispy-forms==1.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
django-phonenumber-field==7.0.1
phonenumbers==8.13.3
